Move Twilio functions to their own folder
